userbot/requirements.txt
pyrogram>=2.0.106
tgcrypto>=1.2.5
pyyaml>=6.0
httpx>=0.27
python-dotenv>=1.0
